fixing a few more things
